// src/data/models/character.ts
export interface ICharacter {
	id: number
	name: string
	status: string
	species: string
	type: string
	gender: string
	origin: {
		name: string
		url: string
	}
	location: {
		name: string
		url: string
	}
	image: string
	episode: string[]
	url: string
	created: string
}

export interface ICharactersResponse {
	info: {
		count: number
		pages: number
		next: string | null
		prev: string | null
	}
	results: ICharacter[]
}

export interface IEpisodeDetails {
	id: number
	name: string
	air_date: string
	episode: string
}